chore: move mocks to own files to reduce conflicts
